KER/dora-openarm-data-collection/nodes/dora-openarm-quitter/src/dora_openarm_quitter/main.py
# Copyright 2026 Enactic, Inc.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

"""Dora node that can quit."""

import dora


def main():
    """Quit input when quit command is received."""
    node = dora.Node()
    for event in node:
        if event["type"] != "INPUT":
            continue

        event_id = event["id"]
        if event_id == "command":
            command = event["value"][0].as_py()
            if command == "quit":
                break
            continue

        node.send_output(event_id, event["value"], event["metadata"])


if __name__ == "__main__":
    main()
